--- src/ratpack/config.groovy
@@ -1,10 +1,0 @@
-import org.ratpackframework.groovy.config.Config
-
-
-(this as Config).with {
-
-	routing.with {
-		reloadable = true
-	}
-
-}
